move client storage classes to their own file
